requirements.txt
python-telegram-bot==20.3
Pillow==10.1.0
pytesseract==0.3.10
python-dotenv==1.0.0